+++ opniLogInferenceService/const.py
