+++ opniLogInferenceService/models/opnilog/opnilog_parser.py
