+++ opniLogInferenceService/start_opnilog_inference.py
