+++ tests/test_train.py
@@ -1,0 +1,4 @@
+# Local
+from opniLogInferenceService import train
+
+train.load_text()
